% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ahccd.R
\name{geomet_ahccd_stns}
\alias{geomet_ahccd_stns}
\title{geomet_ahccd_stns}
\usage{
geomet_ahccd_stns(query)
}
\arguments{
\item{query}{(Optional) List of queryables. This allows for more complicated
queries. See \code{geomet_api_queryables("ahccd-stations")}}
}
\value{
\code{tibble} containing AHCCD station metadata
}
\description{
Get Adjusted Homogenized Canadian Climate Data (AHCCD) station
metadata
}
\examples{
geomet_ahccd_stns()

}
