% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomet_api.R
\name{geomet_api_collections}
\alias{geomet_api_collections}
\title{geomet_api_collections}
\usage{
geomet_api_collections()
}
\value{
\code{tibble} containing collection titles and ids
}
\description{
Lists collections available for query via the MSC Geomet API
}
\examples{
geomet_api_collections()

}
