% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomet_api.R
\name{geomet_api_paginate}
\alias{geomet_api_paginate}
\title{geomet_api_paginate}
\usage{
geomet_api_paginate(req, geometry = TRUE)
}
\arguments{
\item{req}{\code{geomet_api_resp} object}
}
\value{
\code{tibble} containing all paged results
}
\description{
Handles pagination by iterating through results
}
\keyword{internal}
