% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomet_api.R
\name{geomet_api_query}
\alias{geomet_api_query}
\title{geomet_api_query}
\usage{
geomet_api_query(endpoint = "https://api.weather.gc.ca/", path, query = list())
}
\arguments{
\item{endpoint}{The URL endpoint for the API}

\item{path}{The URL path to append to \code{endpoint}}

\item{query}{List of parameters to be passed to query}
}
\value{
\code{geomet_api_resp} object
}
\description{
geomet_api_query
}
\keyword{internal}
