% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomet_api.R
\name{geomet_api_queryables}
\alias{geomet_api_queryables}
\title{geomet_api_queryables}
\usage{
geomet_api_queryables(collection)
}
\arguments{
\item{collection}{Which collection to get queryables for.
See \code{geomet_api_collections}}
}
\value{
\code{tibble} containing possible queryables and their expected type
}
\description{
Provides queryable parameters for API requests
}
\examples{
geomet_api_queryables("hydrometric-stations")

}
