% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate.R
\name{geomet_clim_means}
\alias{geomet_clim_means}
\title{geomet_clim_means}
\usage{
geomet_clim_means(station_number, start_date, end_date, period = "day", query)
}
\arguments{
\item{station_number}{MSC climate station identifier, see
\code{geomet_clim_stns}}

\item{start_date}{Minimum date for values. If blank returns period of record.}

\item{end_date}{Maximum date for values. If blank returns period of record.}

\item{period}{Either "day" for daily means (default) or
"month" for monthly means}

\item{query}{(Optional) List of queryables. This allows for more complicated
queries. Specifying `station_number`, `start_date` and `end_date` arguments
will override `query`.
See \code{geomet_api_queryables("climate-daily")}}
}
\value{
\code{tibble} containing daily mean of all unit values for dates
}
\description{
Get daily means of available climate data
}
\examples{
geomet_clim_means(
  station_number = "1126070",
  start_date = "2016-01-01",
  end_date = "2016-01-02"
)

}
