% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate.R
\name{geomet_clim_normals}
\alias{geomet_clim_normals}
\title{geomet_clim_normals}
\usage{
geomet_clim_normals(station_number, query)
}
\arguments{
\item{station_number}{MSC climate station identifier, see
\code{eccc_clim_stns}}

\item{query}{(Optional) List of queryables. This allows for more complicated
queries. Specifying `station_number` will override `query`.
See \code{geomet_api_queryables("climate-normals")}}
}
\value{
\code{tibble} containing climate normals
}
\description{
Get climate normals for 1981-2010
}
\examples{
geomet_clim_normals(station_number = "1126070")

}
