% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate.R
\name{geomet_clim_stns}
\alias{geomet_clim_stns}
\title{geomet_clim_stns}
\usage{
geomet_clim_stns(query)
}
\arguments{
\item{query}{(Optional) List of queryables. This allows for more complicated
queries. See \code{eccc_queryables("climate-stations")}}
}
\value{
\code{tibble} containing climate station metadata
}
\description{
Get climate station metadata
}
\examples{
geomet_clim_stns()

}
