% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrometric.R
\name{geomet_hydro_annuals}
\alias{geomet_hydro_annuals}
\title{geomet_hydro_annuals}
\usage{
geomet_hydro_annuals(
  station_number,
  start_date,
  end_date,
  type = "stats",
  query
)
}
\arguments{
\item{station_number}{Water Survey of Canada Station identifier.
See \code{geomet_hydro_stns}}

\item{start_date}{Minimum date for values. If blank returns period of record.}

\item{end_date}{Maximum date for values. If blank returns period of record.}

\item{type}{Either "stats" for statistics on daily mean values (default) or
"peaks" for statistics on instantaneous values}

\item{query}{(Optional) List of queryables. This allows for more complicated
queries. Specifying `station_number`, `start_date` and `end_date` arguments
will override `query`.
See \code{geomet_api_queryables("hydrometric-annual-statistics")}}
}
\value{
\code{tibble} containing minimum and maximum for years between
`start_date` and `end_date`
}
\description{
Provides access to minimum and maximum values for daily means
and instantaneous values
}
\examples{
geomet_hydro_annuals(
  station_number = "02HA018",
  start_date = "2019-01-01",
  end_date = "2021-01-01",
  type = "stats"
)

}
