% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrometric.R
\name{geomet_hydro_means}
\alias{geomet_hydro_means}
\title{geomet_hydro_means}
\usage{
geomet_hydro_means(station_number, start_date, end_date, period = "day", query)
}
\arguments{
\item{station_number}{Water Survey of Canada Station identifier.
See \code{geomet_hydro_stns}}

\item{start_date}{Minimum date for values. If blank returns period of record.}

\item{end_date}{Maximum date for values. If blank returns period of record.}

\item{period}{Either "day" for daily means (default) or
"month" for monthly means}

\item{query}{(Optional) List of queryables. This allows for more complicated
queries. Specifying `station_number`, `start_date` and `end_date` arguments
will override `query`.
See \code{geomet_api_queryables("hydrometric-daily-mean")}}
}
\value{
\code{tibble} containing daily mean of all unit values for dates
}
\description{
Get daily mean of water level or flow
}
\examples{
geomet_hydro_means(
  station_number = "04HA001",
  start_date = "2016-01-01",
  end_date = "2016-01-02"
)

}
