% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrometric.R
\name{geomet_hydro_realtime}
\alias{geomet_hydro_realtime}
\title{geomet_hydro_realtime}
\usage{
geomet_hydro_realtime(station_number, query)
}
\arguments{
\item{station_number}{Water Survey of Canada Station identifier.
See \code{geomet_hydro_stns}}

\item{query}{(Optional) List of queryables. This allows for more complicated
queries. Specifying `station_number` will override `query`.
See \code{geomet_api_queryables("hydrometric-realtime")}}
}
\value{
\code{tibble} containing daily mean of all unit values for dates
}
\description{
Get real-time water level and flow data for past 30 days
}
\examples{
geomet_hydro_realtime(station_number = "04HA001")

}
