% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrometric.R
\name{geomet_hydro_stns}
\alias{geomet_hydro_stns}
\title{geomet_hydro_stns}
\usage{
geomet_hydro_stns(query)
}
\arguments{
\item{query}{(Optional) List of queryables.
See \code{geomet_api_queryables("hydrometric-stations")}}
}
\value{
\code{tibble} containing hydrometric station information
}
\description{
Lists hydrometric stations available for query
}
\examples{
geomet_hydro_stns(
  query = list(
    "PROV_TERR_STATE_LOC" = "ON"
  )
)

}
