% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltce.R
\name{geomet_ltce_data}
\alias{geomet_ltce_data}
\title{geomet_ltce_data}
\usage{
geomet_ltce_data(station_number, param, query)
}
\arguments{
\item{station_number}{Virtual station identifier, see \code{geomet_ltce_stns}}

\item{param}{One of "temp", "precip" or "snow"}

\item{query}{(Optional) List of queryables. This allows for more complicated
queries. Specifying `station_number` will override `query`.
See \code{geomet_api_queryables("ltce-temperature")},
\code{geomet_api_queryables("ltce-precipitation")} or
\code{geomet_api_queryables("ltce-snowfall")}}
}
\value{
\code{tibble} containing LTCE values for dates and parameter
}
\description{
Get Long Term Climate Extremes (LTCE) data for virtual climate
stations
}
\examples{
geomet_ltce_data(
  station_number = "VSON99V",
  param = "temp"
)

}
