% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate.R
\name{geomet_swob_realtime}
\alias{geomet_swob_realtime}
\title{geomet_swob_realtime}
\usage{
geomet_swob_realtime(station_number, start_date, end_date, query)
}
\arguments{
\item{station_number}{MSC climate station identifier, see
\code{geomet_clim_stns}}

\item{start_date}{Minimum date for values. Data goes back to a maximum of 30
days}

\item{end_date}{Maximum date for values. Data goes back to a maximum of 30
days}

\item{query}{(Optional) List of queryables. This allows for more complicated
queries. See \code{geomet_api_queryables("swob-realtime")}}
}
\value{
\code{tibble} with SWOB data. Note that this table contains a large
number of columns (>300).
}
\description{
Get real-time surface weather observations (SWOB)
}
\examples{
geomet_swob_realtime(
  station_number = "1192948",
  start_date = Sys.Date() - 2,
  end_date = Sys.Date()
)

}
