% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomet_wcs.R
\name{geomet_wcs_bands}
\alias{geomet_wcs_bands}
\title{geomet_wcs_bands}
\usage{
geomet_wcs_bands(coverage_id, end_point = "geomet")
}
\arguments{
\item{coverage_id}{Product identifier, see `geomet_wcs_capabilities`}

\item{end_point}{Either 'geomet' for weather products or 'geomet-climate'}
}
\value{
List of available bands for product
}
\description{
Gets list of bands for specified product. This is especially useful
  for 'geomet-climate' products where bands are used for time periods
}
