% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomet_wcs.R
\name{geomet_wcs_capabilities}
\alias{geomet_wcs_capabilities}
\title{geomet_wcs_capabilities}
\usage{
geomet_wcs_capabilities(end_point = "geomet")
}
\arguments{
\item{end_point}{Either 'geomet' for weather products or 'geomet-climate'}
}
\value{
Vector of product identifiers
}
\description{
Queries list of available products on GeoMet WCS
}
