% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomet_wcs.R
\name{geomet_wcs_data}
\alias{geomet_wcs_data}
\title{geomet_wcs_data}
\usage{
geomet_wcs_data(coverage_id, query, end_point = "geomet")
}
\arguments{
\item{coverage_id}{Product identifier, see `geomet_wcs_capabilities`}

\item{query}{List of parameters to pass to query, see
\href{https://eccc-msc.github.io/open-data/msc-geomet/web-services_en/#web-coverage-service-wcs}{ECCC Docs}}

\item{end_point}{Either 'geomet' for weather products or 'geomet-climate'}
}
\value{
Path to temporary file with results
}
\description{
Performs query for specified product and query parameters. Saves
  output to temporary file and provides file path. Note that temporary files are
  deleted when the current R session ends.
}
