% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomet_wcs.R
\name{geomet_wcs_query}
\alias{geomet_wcs_query}
\title{Query ECCC GeoMet Web Coverage Service (WCS)}
\usage{
geomet_wcs_query(query, end_point = "geomet", save_to_disk = FALSE)
}
\arguments{
\item{query}{List of query parameters}

\item{end_point}{Either 'geomet' for weather products or 'geomet-climate'
for climate products}

\item{save_to_disk}{TRUE or FALSE, whether to write results to temporary file
instead of returning response object. This can be useful for avoiding
repeating large queries.}
}
\value{
`httr` response or path to file if `save_to_disk`
}
\description{
Query ECCC GeoMet Web Coverage Service (WCS)
}
\keyword{internal}
